requirements.txt
opencv-python
pdf2image
pytesseract
pillow
numpy
pandas
ultralytics
pyyaml  # For config
logging  # Built-in, but note it
reportlab
flask